projects/network_monitor/requirements.txt
psutil==5.9.5
speedtest-cli==2.1.3
